+++ frontend/src/components/features/account/auth/form/LoginForm.tsx
@@ -1,60 +1,47 @@
 'use client'
 
 import type { FieldValues, SubmitHandler } from 'react-hook-form'
+
 import type { z } from 'zod'
+
 import { zodResolver } from '@hookform/resolvers/zod'
 import { useForm } from 'react-hook-form'
 import { useAuth } from '~/hooks/useAuth'
 import { logger } from '~/lib/logger'
-import { signupSchema } from '~/schemas/signup.schema'
-import { Button } from '../button'
+import { loginSchema } from '~/schemas/login.schema'
+import { Button } from '../../../../ui/button'
 import { Form, FormControl, FormDescription, FormField, FormItem, FormLabel, FormMessage } from './Form'
-import { Input } from '../input'
+import { Input } from '../../../../ui/input'
 
 interface LoginFormProps {
   onSuccess?: () => void
 }
 
-export default function SignUpForm({ onSuccess }: LoginFormProps) {
-  const { signup } = useAuth()
+export default function LoginForm({ onSuccess }: LoginFormProps) {
+  const { login } = useAuth()
 
-  const form = useForm<z.infer<typeof signupSchema>>({
-    resolver: zodResolver(signupSchema),
+  const form = useForm<z.infer<typeof loginSchema>>({
+    resolver: zodResolver(loginSchema),
     defaultValues: {
-      name: '',
       email: '',
       password: '',
-      password_confirmation: '',
     },
   })
 
   const onSubmit: SubmitHandler<FieldValues> = async (data) => {
     try {
-      await signup(data.name, data.email, data.password)
+      await login(data.email, data.password)
       if (onSuccess)
         onSuccess()
     }
     catch (error) {
-      logger(error, { tags: { component: 'SignUpForm' } })
+      logger(error, { tags: { component: 'LoginForm' } })
     }
   }
 
   return (
     <Form {...form}>
       <form onSubmit={form.handleSubmit(onSubmit)} className="grid items-start gap-4">
-        <FormField
-          control={form.control}
-          name="name"
-          render={({ field }) => (
-            <FormItem>
-              <FormLabel className="font-bold">ユーザー名</FormLabel>
-              <FormControl>
-                <Input {...field} />
-              </FormControl>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
         <FormField
           control={form.control}
           name="email"
@@ -84,20 +71,7 @@ export default function SignUpForm({ onSuccess }: LoginFormProps) {
             </FormItem>
           )}
         />
-        <FormField
-          control={form.control}
-          name="password_confirmation"
-          render={({ field }) => (
-            <FormItem>
-              <FormLabel className="font-bold">パスワード（確認用）</FormLabel>
-              <FormControl>
-                <Input type="password" {...field} />
-              </FormControl>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-        <Button type="submit" className="h-auto py-3">登録する</Button>
+        <Button type="submit" className="h-auto py-3">ログイン</Button>
       </form>
     </Form>
   )
