+++ frontend/src/components/features/account/auth/form/SignUpForm.tsx
@@ -1,47 +1,60 @@
 'use client'
 
 import type { FieldValues, SubmitHandler } from 'react-hook-form'
-
 import type { z } from 'zod'
-
 import { zodResolver } from '@hookform/resolvers/zod'
 import { useForm } from 'react-hook-form'
 import { useAuth } from '~/hooks/useAuth'
 import { logger } from '~/lib/logger'
-import { loginSchema } from '~/schemas/login.schema'
-import { Button } from '../button'
+import { signupSchema } from '~/schemas/signup.schema'
+import { Button } from '../../../../ui/button'
 import { Form, FormControl, FormDescription, FormField, FormItem, FormLabel, FormMessage } from './Form'
-import { Input } from '../input'
+import { Input } from '../../../../ui/input'
 
 interface LoginFormProps {
   onSuccess?: () => void
 }
 
-export default function LoginForm({ onSuccess }: LoginFormProps) {
-  const { login } = useAuth()
+export default function SignUpForm({ onSuccess }: LoginFormProps) {
+  const { signup } = useAuth()
 
-  const form = useForm<z.infer<typeof loginSchema>>({
-    resolver: zodResolver(loginSchema),
+  const form = useForm<z.infer<typeof signupSchema>>({
+    resolver: zodResolver(signupSchema),
     defaultValues: {
+      name: '',
       email: '',
       password: '',
+      password_confirmation: '',
     },
   })
 
   const onSubmit: SubmitHandler<FieldValues> = async (data) => {
     try {
-      await login(data.email, data.password)
+      await signup(data.name, data.email, data.password)
       if (onSuccess)
         onSuccess()
     }
     catch (error) {
-      logger(error, { tags: { component: 'LoginForm' } })
+      logger(error, { tags: { component: 'SignUpForm' } })
     }
   }
 
   return (
     <Form {...form}>
       <form onSubmit={form.handleSubmit(onSubmit)} className="grid items-start gap-4">
+        <FormField
+          control={form.control}
+          name="name"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel className="font-bold">ユーザー名</FormLabel>
+              <FormControl>
+                <Input {...field} />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
         <FormField
           control={form.control}
           name="email"
@@ -71,7 +84,20 @@ export default function LoginForm({ onSuccess }: LoginFormProps) {
             </FormItem>
           )}
         />
-        <Button type="submit" className="h-auto py-3">ログイン</Button>
+        <FormField
+          control={form.control}
+          name="password_confirmation"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel className="font-bold">パスワード（確認用）</FormLabel>
+              <FormControl>
+                <Input type="password" {...field} />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+        <Button type="submit" className="h-auto py-3">登録する</Button>
       </form>
     </Form>
   )
